Refactor: do not export functions if necessary
